requirements.txt
alembic==1.4.3
Babel==2.8.0
bcrypt==3.2.0
blinker==1.4
cffi==1.14.3
click==7.1.2
cryptography==3.1.1
dnspython==2.0.0
dominate==2.5.2
email-validator==1.1.1
et-xmlfile==1.0.1
Flask==1.1.2
Flask-Admin==1.5.6
Flask-Babel==2.0.0
Flask-Bootstrap==3.3.7.1
Flask-Excel==0.0.7
Flask-FontAwesome==0.1.5
Flask-Login==0.5.0
Flask-Mail==0.9.1
Flask-Markdown==0.3
Flask-Migrate==2.5.3
Flask-Script==2.0.6
Flask-SQLAlchemy==2.4.4
Flask-Table==0.5.0
Flask-User==1.0.2.2
Flask-WTF==0.14.3
idna==2.10
importlib-metadata==2.0.0
itsdangerous==1.1.0
jdcal==1.4.1
Jinja2==2.11.2
lml==0.0.9
Mako==1.1.3
Markdown==3.3
MarkupSafe==1.1.1
openpyxl==3.0.5
passlib==1.7.4
psycopg2==2.8.6
pycparser==2.20
pyexcel==0.6.5
pyexcel-io==0.6.2
pyexcel-webio==0.1.4
python-dateutil==2.8.1
python-dotenv==0.14.0
python-editor==1.0.4
pytz==2020.1
six==1.15.0
SQLAlchemy==1.3.19
texttable==1.6.3
visitor==0.1.3
Werkzeug==1.0.1
WTForms==2.3.3
zipp==3.3.0
